fixed println and printf issue and refactored code
